added role tests

+++ src/modules/panel/services/role.js
@@ -1,6 +1,7 @@
 import RoleData from '../data/role.js';
-import permissionMiddleware from '../../../utils/permission-middleware.js';
-import TokenHelper from '../../../utils/token-helper.js';
+import Errors from '../constants/error.js';
+import ErrorHelper from '../../../utils/error-helper.js';
+
 class RoleService {
     static async createRole({ userId, name, description, color, authEndpoints }) {
         const role = await RoleData.createRole({ userId, name, description, color, authEndpoints });
@@ -8,13 +9,21 @@ class RoleService {
     }
 
     static async updateRole({ userId, roleId, name, description, color, authEndpoints }) {
-        const role = await RoleData.updateRole({ userId, roleId, name, description, color, authEndpoints });
-        return role;
+        const role = await RoleData.getRoleById({ roleId });
+        if (!role) {
+            throw new ErrorHelper(Errors.ROLE_ERROR.message, Errors.ROLE_ERROR.statusCode)
+        }
+        const updatedRole = await RoleData.updateRole({ userId, roleId, name, description, color, authEndpoints });
+        return updatedRole;
     }
 
     static async deleteRole({ userId, roleId }) {
-        const role = await RoleData.deleteRole({ userId, roleId });
-        return role;
+        const role = await RoleData.getRoleById({ roleId });
+        if (!role) {
+            throw new ErrorHelper(Errors.ROLE_ERROR.message, Errors.ROLE_ERROR.statusCode)
+        }
+        const deletedRole = await RoleData.deleteRole({ userId, roleId });
+        return deletedRole;
     }
 
     static async getRoles({ page, limit }) {
